Move WriteRuneToFile to font package
